backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
sentence-transformers>=2.2.2
numpy>=2.0.0
requests>=2.32.0
aiofiles>=23.2.1
PyPDF2>=3.0.1
pytesseract>=0.3.10
Pillow>=10.1.0
pdf2image>=1.16.3
openai>=1.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
qdrant-client>=1.7.0
loguru>=0.7.2
slowapi>=0.1.9
pyjwt>=2.8.0
